pyproject.toml
[tool.poetry]
name = "fast-zero"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
psycopg = {extras = ["binary"], version = "^3.2.1"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
taskipy = "^1.13.0"
fastapi = "^0.111.0"
ipdb = "^0.13.13"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"
testcontainers = "^4.8.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run = 'clear ; fastapi dev fast_zero/app.py'
format = 'ruff check . --fix && ruff format .'

pre_test = 'clear; task lint'
test = 'pytest --cov=fast_zero -vv'
post_test = 'coverage html'

lint = 'ruff check . ; ruff check . --diff'

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--pdbcls=IPython.terminal.debugger:Pdb"
# addopts = '-p no:warnings --pdbcls=IPython.terminal.debugger:Pdb'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
